[Cache][Messenger] Restore compatibility with ext-redis 5
Dropping the "ext-redis: <6.1" conflict makes symfony/cache and
symfony/redis-messenger installable again on distributions that only
ship ext-redis 5 (Debian Bookworm, Ubuntu 24.04), which is where the
7.4 security releases became unreachable.

The version-specific proxies are restored the way 6.4 still ships them:
RedisProxy/RedisClusterProxy alias to the 5 or 6 flavor depending on
phpversion('redis'), and the 6.1 signature differences move back to
Redis61ProxyTrait/RedisCluster61ProxyTrait next to the existing 6.2 and
6.3 traits.
