+++ src/Symfony/Component/Cache/composer.json
@@ -43,7 +43,6 @@
         "symfony/var-dumper": "^6.4|^7.0|^8.0"
     },
     "conflict": {
-        "ext-redis": "<6.1",
         "ext-relay": "<0.12.1",
         "doctrine/dbal": "<3.6",
         "symfony/dependency-injection": "<6.4",
